Implement row division.
This is a scary function, since if the user
decides to divide a row by a scalar that isn't
a factor of one of the elements in the row,
then the matrix itself is compromised. At
some point, we'll have to use fraction structs
or instead switch to doubles as the matrix type.